+++ src/components/app.tsx
@@ -1,5 +1,4 @@
 import {MainPage} from './pages/main-page/main-page';
-import {CommonProps} from '../mocks/types';
 import {BrowserRouter, Route, Routes, Outlet } from 'react-router-dom';
 import {SignIn} from './pages/sing-in-page/sign-in';
 import {MyList} from './pages/my-list-page/my-list';
@@ -8,6 +7,7 @@ import {MoviePage} from './pages/movie-page/movie-page';
 import {AddReview} from './pages/movie-page/add-review';
 import {ErrorPage} from './pages/error-page';
 import {PrivateRoute} from './private-route';
+import {CommonProps} from './types/common';
 
 export const App = (props: CommonProps)=> (
   <BrowserRouter>
