+++ src/components/pages/main-page/film-list.tsx
@@ -1,18 +1,18 @@
 import {FilmCard} from './film-card';
 import {useState} from 'react';
-import {FilmCards} from '../../../mocks/types';
+import {FilmCards} from '../../types/film';
 
 export const FilmList = (props: {films: FilmCards[]; onlyImage: boolean }) => {
-  const [activeFilm, setActiveCard] = useState<FilmCards | null>(null);
+  const [activeFilmId, setActiveCardId] = useState<number | null>(null);
   return (
     <div className="catalog__films-list">
       {props.films.map((film) =>
         (
           <FilmCard
-            isActive={!props.onlyImage && activeFilm?.id === film.id}
-            key={film.id} {...film}
-            onMouseOver={() => setActiveCard(film)}
-            onMouseOut={() => setActiveCard(null)}
+            isActive={!props.onlyImage && activeFilmId === film.id}
+            key={film.id} filmData={film}
+            onMouseOver={() => setActiveCardId(film.id)}
+            onMouseOut={() => setActiveCardId(null)}
           />
         )
       )}
