+++ src/components/pages/movie-page/tabs/details.tsx
@@ -1,4 +1,4 @@
-import {DetailsProps} from '../../../../mocks/tabProps';
+import {DetailsProps} from '../../../types/tabs';
 
 export const Details = (props: DetailsProps) => (
   <div className="film-card__text film-card__row">
