+++ src/components/types/film.ts
@@ -1,0 +1,9 @@
+export type FilmCards = {
+  id: number;
+  title: string;
+  previewImage: string;
+  previewVideoLink: string;
+  genre?: string;
+  year?: number;
+  bgImage?: string;
+};
