+++ src/components/types/player.ts
@@ -1,0 +1,6 @@
+export type PlayerProps = {
+  src: string;
+  poster: string;
+  time: string;
+  name: string;
+};
